# api/docs/swagger.yaml
basePath: /
definitions:
...
  errors.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      path:
        type: string
      request_id:
        type: string
      retry_after:
        type: integer
      timestamp:
        type: string
    type: object
...
  errors.Meta:
    properties:
      links:
        $ref: '#/definitions/errors.Links'
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      timestamp:
        type: string
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  errors.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/errors.ErrorInfo'
      meta:
        $ref: '#/definitions/errors.Meta'
      success:
        type: boolean
    type: object
...
  imoveis.Anexo:
    properties:
      canPublish:
        type: boolean
      created_at:
        type: string
      empreendimento_id:
        type: integer
      id:
        type: integer
      image:
        type: boolean
      imovel_id:
        type: integer
      isExternalUrl:
        type: boolean
      nome:
        type: string
      path:
        type: string
      planta_id:
        type: integer
      tamanho:
        type: integer
      tipo:
        type: string
      updated_at:
        type: string
      url:
        type: string
      video:
        type: boolean
    type: object
  imoveis.AnexoResponse:
    properties:
      canPublish:
        type: boolean
      created_at:
        type: string
      id:
        type: integer
      image:
        type: boolean
      isExternalUrl:
        type: boolean
      nome:
        type: string
      path:
        type: string
      tamanho:
        type: integer
      tipo:
        type: string
      updated_at:
        type: string
      url:
        type: string
      video:
        type: boolean
    type: object
  imoveis.CaracteristicaResponse:
    properties:
      categoria_id:
        type: integer
      categoria_nome:
        type: string
      created_at:
        type: string
      id:
        type: integer
      nome:
        type: string
      updated_at:
        type: string
    type: object
  imoveis.CorretorPrincipalResponse:
    properties:
      bairrosAtuacao:
        items:
          type: string
        type: array
      email:
        type: string
      foto:
        $ref: '#/definitions/imoveis.AnexoResponse'
      id:
        type: integer
      idiomas:
        items:
          type: string
        type: array
      nome:
        type: string
      organizacao:
        $ref: '#/definitions/imoveis.OrganizacaoResponse'
      whatsapp:
        type: string
    type: object
  imoveis.CreateImovelRequest:
    properties:
      caracteristicas:
        items:
          type: integer
        type: array
      codigo:
        maxLength: 50
        minLength: 1
        type: string
      condominio:
        minimum: 0
        type: number
      corretor_principal_id:
        type: integer
      descricao:
        maxLength: 5000
        minLength: 10
        type: string
      empreendimento_id:
        type: integer
      endereco_id:
        description: Relations
        type: integer
      finalidade:
        enum:
        - RESIDENTIAL
        - COMERCIAL
        - MISTO
        type: string
      id_integracao:
        type: string
      inscricaoIPTU:
        maxLength: 50
        type: string
      iptu:
        minimum: 0
        type: number
      metragem:
        type: number
      numAndar:
        type: integer
      numBanheiros:
        minimum: 0
        type: integer
      numQuartos:
        minimum: 0
        type: integer
      numSuites:
        minimum: 0
        type: integer
      numVagas:
        minimum: 0
        type: integer
      objetivo:
        enum:
        - VENDER
        - ALUGAR
        type: string
      pacote_id:
        type: integer
      planta_id:
        type: integer
      preco_aluguel_id:
        type: integer
      preco_venda_id:
        type: integer
      tipo:
        enum:
        - APARTAMENTO
        - CASA
        - COMERCIAL
        - SALA_COMERCIAL
        - TERRENO
        - GALPAO
        type: string
      titulo:
        maxLength: 255
        minLength: 3
        type: string
      unidade:
        maxLength: 20
        type: string
    required:
    - codigo
    - descricao
    - endereco_id
    - finalidade
    - metragem
    - objetivo
    - tipo
    - titulo
    type: object
  imoveis.EmpreendimentoResponse:
    properties:
      anexos:
        items:
          $ref: '#/definitions/imoveis.AnexoResponse'
        type: array
      caracteristicas:
        items:
          $ref: '#/definitions/imoveis.CaracteristicaResponse'
        type: array
      created_at:
        type: string
      data_entrega:
        type: string
      descricao:
        type: string
      endereco:
        $ref: '#/definitions/imoveis.EnderecoResponse'
      etapa_lancamento:
        type: string
      finalidade:
        type: string
      id:
        type: integer
      localizacao:
        type: string
      plantas:
        items:
          $ref: '#/definitions/imoveis.PlantaResponse'
        type: array
      status:
        type: string
      tipo:
        type: string
      titulo:
        type: string
      torres:
        items:
          $ref: '#/definitions/imoveis.TorresResponse'
        type: array
      updated_at:
        type: string
    type: object
  imoveis.EnderecoResponse:
    properties:
      bairro:
        type: string
      cep:
        type: string
      cidade:
        type: string
      estado:
        type: string
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      numero:
        type: integer
      rua:
        type: string
    type: object
  imoveis.ImovelListResponse:
    properties:
      hasNext:
        type: boolean
      hasPrev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      pages:
        type: integer
      results:
        items:
          $ref: '#/definitions/imoveis.ImovelResponse'
        type: array
      total:
        type: integer
    type: object
  imoveis.ImovelResponse:
    properties:
      anexos:
        items:
          $ref: '#/definitions/imoveis.AnexoResponse'
        type: array
      caracteristicas:
        items:
          $ref: '#/definitions/imoveis.CaracteristicaResponse'
        type: array
      closed:
        type: boolean
      codigo:
        type: string
      condominio:
        type: number
      corretorPrincipal:
        $ref: '#/definitions/imoveis.CorretorPrincipalResponse'
      created_at:
        type: string
      descricao:
        type: string
      empreendimento:
        $ref: '#/definitions/imoveis.EmpreendimentoResponse'
      endereco:
        allOf:
        - $ref: '#/definitions/imoveis.EnderecoResponse'
        description: Relations
      finalidade:
        type: string
      id:
        type: integer
      id_integracao:
        type: string
      inscricaoIPTU:
        type: string
      iptu:
        type: number
      metragem:
        type: number
      numAndar:
        type: integer
      numBanheiros:
        type: integer
      numQuartos:
        type: integer
      numSuites:
        type: integer
      numVagas:
        type: integer
      objetivo:
        type: string
      pacote:
        $ref: '#/definitions/imoveis.PacoteResponse'
      planta:
        $ref: '#/definitions/imoveis.PlantaResponse'
      precoAluguel:
        $ref: '#/definitions/imoveis.PrecoAluguelResponse'
      precoVenda:
        $ref: '#/definitions/imoveis.PrecoVendaResponse'
      published:
        type: boolean
      seqCodigo:
        type: integer
      status:
        description: Metadata
        type: string
      tipo:
        type: string
      titulo:
        type: string
      unidade:
        type: string
      updated_at:
        type: string
      visualizacoes:
        type: integer
    type: object
...
  imoveis.PacoteResponse:
    properties:
      created_at:
        type: string
      descricao:
        type: string
      em_destaque:
        type: boolean
      exclusivo:
        type: boolean
      id:
        type: integer
      titulo:
        type: string
      updated_at:
        type: string
    type: object
  imoveis.PlantaResponse:
    properties:
      anexos:
        items:
          $ref: '#/definitions/imoveis.AnexoResponse'
        type: array
      created_at:
        type: string
      id:
        type: integer
      metragem:
        type: number
      nome:
        type: string
      updated_at:
        type: string
    type: object
...
  imoveis.UpdateImovelRequest:
    properties:
      caracteristicas:
        items:
          type: integer
        type: array
      closed:
        type: boolean
      codigo:
        maxLength: 50
        minLength: 1
        type: string
      condominio:
        minimum: 0
        type: number
      corretor_principal_id:
        type: integer
      descricao:
        maxLength: 5000
        minLength: 10
        type: string
      empreendimento_id:
        type: integer
      endereco_id:
        description: Relations
        type: integer
      finalidade:
        enum:
        - RESIDENTIAL
        - COMERCIAL
        - MISTO
        type: string
      inscricaoIPTU:
        maxLength: 50
        type: string
      iptu:
        minimum: 0
        type: number
      metragem:
        type: number
      numAndar:
        type: integer
      numBanheiros:
        minimum: 0
        type: integer
      numQuartos:
        minimum: 0
        type: integer
      numSuites:
        minimum: 0
        type: integer
      numVagas:
        minimum: 0
        type: integer
      objetivo:
        enum:
        - VENDER
        - ALUGAR
        type: string
      pacote_id:
        type: integer
      planta_id:
        type: integer
      preco_aluguel_id:
        type: integer
      preco_venda_id:
        type: integer
      published:
        type: boolean
      status:
        enum:
        - PUBLICADO
        - EM_EDICAO
        - ARQUIVADO
        type: string
      tipo:
        enum:
        - APARTAMENTO
        - CASA
        - COMERCIAL
        - SALA_COMERCIAL
        - TERRENO
        - GALPAO
        type: string
      titulo:
        maxLength: 255
        minLength: 3
        type: string
      unidade:
        maxLength: 20
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A production-ready REST API boilerplate in Go with JWT authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: TRIIIO API Boilerplate
  version: "1.0"
paths:
...
  /api/v1/imoveis:
    get:
      consumes:
      - application/json
      description: Get paginated list of properties with filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Property code (partial match)
        in: query
        name: codigo
        type: string
      - description: Property type (APARTAMENTO, CASA, COMERCIAL, SALA_COMERCIAL,
          TERRENO, GALPAO)
        in: query
        name: tipo
        type: string
      - description: Property objective (VENDER, ALUGAR)
        in: query
        name: objetivo
        type: string
      - description: Property purpose (RESIDENTIAL, COMERCIAL, MISTO)
        in: query
        name: finalidade
        type: string
      - description: Property status (PUBLICADO, EM_EDICAO, ARQUIVADO)
        in: query
        name: status
        type: string
      - description: Published status
        in: query
        name: published
        type: boolean
      - description: Minimum price
        in: query
        name: min_preco
        type: number
      - description: Maximum price
        in: query
        name: max_preco
        type: number
      - description: Minimum square meters
        in: query
        name: min_metragem
        type: number
      - description: Maximum square meters
        in: query
        name: max_metragem
        type: number
      - description: Street name (partial match)
        in: query
        name: rua
        type: string
      - description: City name (partial match)
        in: query
        name: cidade
        type: string
      - description: Neighborhood name (partial match)
        in: query
        name: bairro
        type: string
      - description: Minimum number of bedrooms
        in: query
        name: num_quartos
        type: integer
      - description: Minimum number of bathrooms
        in: query
        name: num_banheiros
        type: integer
      - description: Minimum number of parking spaces
        in: query
        name: num_garagens
        type: integer
      - description: Development ID
        in: query
        name: empreendimento_id
        type: integer
      - default: created_at
        description: Sort field (created_at, updated_at, preco, titulo, metragem)
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  $ref: '#/definitions/imoveis.ImovelListResponse'
                success:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      summary: List properties
      tags:
      - imoveis
    post:
      consumes:
      - application/json
      description: Create a new property
      parameters:
      - description: Property creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/imoveis.CreateImovelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  $ref: '#/definitions/imoveis.ImovelResponse'
                success:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Create a new property
      tags:
      - imoveis
  /api/v1/imoveis/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a property
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Delete a property
      tags:
      - imoveis
    get:
      consumes:
      - application/json
      description: Get a property by its ID
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  $ref: '#/definitions/imoveis.ImovelResponse'
                success:
                  type: boolean
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      summary: Get property by ID
      tags:
      - imoveis
    put:
      consumes:
      - application/json
      description: Update an existing property
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      - description: Property update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/imoveis.UpdateImovelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  $ref: '#/definitions/imoveis.ImovelResponse'
                success:
                  type: boolean
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Update a property
      tags:
      - imoveis
  /api/v1/imoveis/{id}/anexos:
    get:
      consumes:
      - application/json
      description: Get all attachments for a property
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/imoveis.AnexoResponse'
                  type: array
                success:
                  type: boolean
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      summary: Get property attachments
      tags:
      - imoveis
    post:
      consumes:
      - application/json
      description: Add an image or document attachment to a property
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      - description: Attachment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/imoveis.Anexo'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Add attachment to property
      tags:
      - imoveis
  /api/v1/imoveis/{id}/caracteristicas:
    get:
      consumes:
      - application/json
      description: Get all characteristics for a property
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/imoveis.CaracteristicaResponse'
                  type: array
                success:
                  type: boolean
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      summary: Get property characteristics
      tags:
      - imoveis
    post:
      consumes:
      - application/json
      description: Add multiple characteristics to a property
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      - description: Characteristics IDs
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            items:
              type: integer
            type: array
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Add characteristics to property
      tags:
      - imoveis
  /api/v1/imoveis/import:
    post:
      consumes:
      - application/json
      description: Import all published properties from dev-api-backend.pi8.com.br.
        Uses upsert logic - creates new properties and updates existing ones based
        on id_integracao mapping. Existing properties are detected via id_integracao
        field and updated with latest data. Attachments are deduplicated by URL.
      produces:
      - application/json
      responses:
        "200":
          description: Import completed with statistics (created, updated, failed
            counts)
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Import properties from external API
      tags:
      - imoveis
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"